+++ CHANGES.md
@@ -4,6 +4,11 @@
 - major refactor of the code and dependencies
 - new configuration format (YAML)
 - better API
+- add ability to disable common names using CCD
+- get a list of disabled common names
+- stricter input validation
+- add some unit tests
+- do not require authentication to refresh CRL
 
 ## 1.1.0 (2015-12-22)
 - add server info API call